requirements.txt
tensorflow>=2.13
keras>=2.13
flask>=3.0
numpy>=1.24
opencv-python>=4.9
pillow>=10.0
matplotlib>=3.8
scikit-learn>=1.4
gunicorn>=21.2
gradio>=4.0
